Update homebrew installation command
The command to install via homebrew didn't work for me. Presumably this changed in a recent version of homebrew.

This is the error I received:
```
$ brew install coursier/formulas/coursier

Error: coursier/formulas/coursier is a head-only formula
Install with `brew install --HEAD coursier/formulas/coursier`
```

So this change just incorporates the fix it suggests, which works for me.