Added simple ZSH completions, and updated README with install details
These completions were generated by copy/pasting details from the
version 1.0.0 '--help' output, and had a small amount of human input.
The completions are by no means perfect, and there is no intelligent
completion of parameter values (except `--mode`, which was trivial).

I hope that completions prove to be useful, and can form a basis that
other users of coursier can build upon, to improve the interactivity of
the command-line client.