Allow to use several caches when fetching artifacts
Useful to silently pick local things via a first cache instance, before
a second one downloads the rest.

The number of artifacts passed as size hint to the second cache logger
is then closer to the number of artifacts that are going to actually be
downloaded. (Whereas with a single cache, that number is roughly
the number of artifacts picked from the cache + number of artifacts to
be downloaded.)