Allow to generate hybrid standalone / assembly launchers (#1233)
These launchers are still started by the code in the bootstrap-launcher
module. The JARs that should have been embedded in the launcher and
loaded by a first class loader are instead put in the launcher like in
an assembly (their content as is in the launcher zip).

That allows to access those JARs via a simple file, rather than via a
zip-in-a-zip URL. Which can sometimes allow for faster things for users.