Parse homepage attribute in ivy.xml
This extracts the [`homepage`](https://ant.apache.org/ivy/history/latest-milestone/ivyfile/description.html)
attribute from the `description` element in `ivy.xml` files and sets it
as `homePage` in `Info`.

My goal is to use this in Scala Steward for linking to home pages and
release notes of sbt plugins.