Allow to make some properties optional
That is, only set them if they're not already set.

For example, when adding a property like "foo?": "value" to an app
descriptor, "foo" will be set to "value" only if it wasn't set by the
user. That allows users to override a property value when launching an
app by passing '-J-Dfoo=custom-value' to it.