Delete lockfile explicitly
This partially reverts the changes introduced in 344f6e73c10cf161c25a5878d6db9cce470e02b0. Specifically, it reverts back to explicitly deleting lockfiles from caches, rather than relying on the DELETE_ON_CLOSE channel option.

It solves a concurrent download bug in mill and unblocks this PR https://github.com/lihaoyi/mill/pull/858. I'm not exactly sure why this change fixes the issue, but this is what I found after many hours of bisecting. Note that it is not sufficient to explicitly delete the file in the finally block, but that the DELETE_ON_CLOSE must not be used as it somehow interferes.