Fix Scala version enforcement
Fixes #2054

Problem
-------
Currently `forceScalaVersionOpt` option, which by default is set to true
when a Scala version is present, conflates the notion of overriding the
Scala standard modules (such as scala-library, scala-reflect) and
overriding the cross building full suffix of any libraries.

This ignores the user's specification of dependency such as
`"org.typelevel" % "macro-compat_2.13.0-RC2" % "1.1.1"`
and substitutes it with `"macro-compat_2.13.5"` etc.

Solution
--------
Create a new option specifically to override the full suffix, and make
it opt-in.
Scala modules should continue to enjoy the Scala version overriding.