Be fine with Mill-generated manifests
Mill doesn't add Implementation-Vendor-Id and Specification-Title in the
manifests it generates, which the main class detection code uses to pick
a main class.

These changes default to the main class from the first JAR in the class
path (which has been determistic for some time now, and corresponds to the
JAR of the first specified dependency), no matter its
Implementation-Vendor-Id and Specification-Title attributes.