Introduce `VersionOverride`
A version override is a way to override the app descriptor based on the app version.
This change is introduced to handle both scala-compiler and scala3-compiler
under the same app descriptor:

{
  "mainClass": "dotty.tools.dotc.Main",
  "repositories": [
    "central"
  ],
  "dependencies": [
    "org.scala-lang:scala3-compiler_3:latest.stable"
  ],
  "properties": {
    "scala.usejavacp": "true"
  },
  "versionOverrides": [
    {
      "versionInterval": "(,2.max]",
      "mainClass": "scala.tools.nsc.MainGenericRunner",
      "dependencies": [
        "org.scala-lang:scala-compiler:latest.stable"
      ]
    }
  ]
}

In that case, if no version is specified, Coursier will install the latest
stable Scala 3 version. However, if the user specifies a version below 3,
Coursier will install the Scala 2 compiler whose maven coordinates and
main class are different.