fix: make sure JAVA_HOME is used in bootstrapped file (#2441)
I came across this for some Metals users that use coursier to bootstrap
a metals executable. The executable just calls `java` and doesn't
respect `JAVA_HOME`. This change makes sure that in the shell script if
`JAVA_HOME` is set, it's used and if not we fall back to `java`.

Closes #2048