Support `--architecture` flag for Java commands
Previously, it was not possible to use `coursier java-home` to download
an x86_64 architecture JDK on an Apple M1 computer. This may be a
niche use-case, but we can't use any existing aarch64 JDK 8
distributions in the https://github.com/sourcegraph/scip-java build
because none of them give access to the `com.sun` APIs as far as I can
tell. With this commit, we can use `--architecture amd64 --jvm 8` to
populate the boot classpath of our javac compiler plugin so that
`com.sun` APIs are available during compile time.