`intern` strings in `JsonElem`
I ran into an issue where `coursier fetch --json-output-file ...` was
failing with `java.lang.OutOfMemoryError`. This failure persisted even
with heap sizes of 8 GB. Analysis of a heap dump showed that 66% of the
heap consisted of duplicate strings. In my particular case, 9% of the
heap was taken up by 149,403 copies of the string
`"org.scala-lang:scala-library:2.12.12"`, and another 9% was copies of
the string `"org.scala-lang:scala-library:2.12.8"`. It appeared that
most of these strings were held by the `reconciledVersionStr` and
`requestedVersionStr` fields of `JsonElem`.

This PR adds `intern` calls to the allocations of `reconciledVersion`,
`reconciledVersionStr`, and `requestedVersionStr` on `JsonElem`, causing
these string allocations to be effectively deduplicated in to the JVM
string pool. After this change, my same fetch command was able to complete
running with a much smaller heap, `-Xmx256m`.

This PR also adds `lazy` to `reconciledVersionStr`, as the eager evaluation
of `reconciledVersionStr` effectively forced `reconciledVersion` to be
non-lazy as well.