drop jabba from pre-installed JVM indexes
jabba index has not been updated for a while therefore there is no need
to keep extra code to support it. Moreover, dropping it fixes
https://github.com/coursier/coursier/issues/2487 because downloaded
index was modified by `JvmIndex.jabbaIndexGraalvmJava8Hack` function
with no respect to index source - so not only `jabba` index was modified
by it. For example, for coursier index
https://github.com/coursier/jvm-index it messed up graalvm versions
badly and caused the lost of latest graalvm versions.