Always build Linux binaries from the controlled Docker image if we are in the CI
We used to switch to the Docker environment in case the host environment’s architecture was `x86_64`, but `ubuntu:latest` is currently `amd64`. Now we always use the Docker environment to build the Linux launchers’ binaries when we are in the CI to make sure we don’t depend on the current state of the `ubuntu` image provided by GitHub.