Preserve multi-release flag in JARs with shading (#3442)
* Preserve multi-release flag in JARs with shading

Before these changes, if a JAR (like coursier-cache) was shading multi-release
JARs, the multi-release flag wasn't added in its manifest, which is required
for classes for newer JVM versions in multi-release JARs to be loaded. Because
of this, new JDK foreign API stuff used on Windows to compute the cache
directory weren't being used on JDK >= 23.

* Fix Windows build if cs isn't cs.exe in PATH