Add --reference-url option to 'cs get' command
This allows to mimick the Artifact instances generated by the coursier
resolver. coursier adds POMs as "reference artifacts" to JARs and all.
Later on, if the cache gets a not-found error for an artifact, but its
reference artifact is in cache, it persists the error on disk, assuming
the artifact will not appear later on on the server (if the POM is on
the server, we assume all artifacts are alongside him).